+++ mirror-box/__init__.py
@@ -1,7 +1,7 @@
 bl_info = {
 	"name":"Mirror Box",
 	"description":"Generate a mirrored, subdivided version of the starting cube",
-	"version":(0,1,0),
+	"version":(0,1,1),
 	"blender":(2,78,0),
 	"support":"TESTING",
 	"category":"Objects",
@@ -9,34 +9,29 @@ bl_info = {
 }
 import bpy
 
-
-
-class MirrorBox(bpy.types.Operator):
-	bl_idname="object.mirrorbox"
-	bl_label = "Add MirrorBox"
-	bl_options = {"REGISTER", "UNDO"}
-
-	def execute(self, context):
+class MirrorBoxInfo:
+	def __init__(self, radius=1.0):
 		#Right half of a once subdivided cube
-		verts = [(0.0, -1.0, 1.0),	#0
-			(1.0, -1.0, 1.0),		#1
-			(1.0, 0.0, 1.0),		#2
-			(0.0, 0.0, 1.0),		#3
-			(0.0, 1.0, 1.0),		#4
-			(1.0, 1.0, 1.0),		#5
-			(0.0, -1.0, 0.0),		#6
-			(1.0, -1.0, 0.0),		#7
-			(1.0, 0.0, 0.0),		#8
-			(1.0, 1.0, 0.0),		#9
-			(0.0, 1.0, 0.0),		#10
-			(0.0, -1.0, -1.0),		#11
-			(1.0, -1.0, -1.0),		#12
-			(1.0, 0.0, -1.0),		#13
-			(1.0, 1.0, -1.0),		#14
-			(0.0, 1.0, -1.0),		#15
-			(0.0, 0.0, -1.0)]		#16
-		edges = []
-		faces = [(0,1,2,3),
+		self.verts = [
+			(0.0, -radius, radius),			#0
+			(radius, -radius, radius),		#1
+			(radius, 0.0, radius),			#2
+			(0.0, 0.0, radius),				#3
+			(0.0, radius, radius),			#4
+			(radius, radius, radius),		#5
+			(0.0, -radius, 0.0),			#6
+			(radius, -radius, 0.0),			#7
+			(radius, 0.0, 0.0),				#8
+			(radius, radius, 0.0),			#9
+			(0.0, radius, 0.0),				#10
+			(0.0, -radius, -radius),		#11
+			(radius, -radius, -radius),		#12
+			(radius, 0.0, -radius),			#13
+			(radius, radius, -radius),		#14
+			(0.0, radius, -radius),			#15
+			(0.0, 0.0, -radius)]			#16
+		self.edges = []
+		self.faces = [(0,1,2,3),
 			(5, 4, 3, 2),
 			(1, 0, 6, 7),
 			(7, 8, 2, 1),
@@ -49,8 +44,18 @@ class MirrorBox(bpy.types.Operator):
 			(16, 13, 12, 11),
 			(16, 15, 14, 13)]
 
+
+class MirrorBoxOperator(bpy.types.Operator):
+	bl_idname="object.mirrorbox"
+	bl_label = "Add MirrorBox"
+	bl_options = {"REGISTER", "UNDO"}
+
+	def execute(self, context):
+		
+
 		mesh_data = bpy.data.meshes.new("mirror_box_mesh_data")
-		mesh_data.from_pydata(verts, edges, faces)
+		meshInfo = MirrorBoxInfo()
+		mesh_data.from_pydata(meshInfo.verts, meshInfo.edges, meshInfo.faces)
 		mesh_data.update()
 
 		obj = bpy.data.objects.new("MrBox", mesh_data)
@@ -64,10 +69,10 @@ class MirrorBox(bpy.types.Operator):
 		return {'FINISHED'}
 
 def register():
-	bpy.utils.register_class(MirrorBox)
+	bpy.utils.register_class(MirrorBoxOperator)
 
 def unregister():
-	bpy.utils.unregister_class(MirrorBox)
+	bpy.utils.unregister_class(MirrorBoxOperator)
 
 if __name__ == "__main__":
 	register()